refactor: replace contants.ts for variables.ts
