+++ src/components/checkout/CheckoutPage.tsx
@@ -1,23 +1,20 @@
 import { BackButton } from "@/components/shared/buttons/BackButton"
 import { CheckoutSection } from "./components/CheckoutSection"
-
-import "./styles.scss"
 import { ConfirmationModal } from "./components/checkoutModal"
+import "./styles.scss"
 
 export function CheckoutPageComponent() {
     return (
-        <>
-            <div className="checkout-page-container">
-                <div className="utils-wrapper">
-                    <div className="utils-wrapper__inner">
-                        <BackButton />
-                    </div>
+        <div className="checkout-page-container">
+            <div className="utils-wrapper">
+                <div className="utils-wrapper__inner">
+                    <BackButton />
                 </div>
-                <CheckoutSection />
-                <ConfirmationModal.Root>
-                    <ConfirmationModal.Items />
-                </ConfirmationModal.Root>
             </div>
-        </>
+            <CheckoutSection />
+            <ConfirmationModal.Root>
+                <ConfirmationModal.Items />
+            </ConfirmationModal.Root>
+        </div>
     )
 }
