+++ src/components/checkout/helpers/variables.ts
